+++ StepStone/Framework/Graphics.h
@@ -73,7 +73,7 @@ namespace Graphics
 	//This draws the texture flipped, so perfect for a directly loaded png!
 	void drawImage(unsigned int width, unsigned int height, GLfloat scaledWidth, GLfloat scaledHeight);
 	void drawImage(int x, int y, int width, int height, GLfloat scaledWidth, GLfloat scaledHeight);
-	void drawImage(unsigned int x, unsigned int y, float width, float height, GLfloat scaledWidth, GLfloat scaledHeight);
+	void drawImage(float x, float y, float width, float height, GLfloat scaledWidth, GLfloat scaledHeight);
 	void drawIntroImage(unsigned int width, unsigned int height, GLfloat scaledWidth, GLfloat scaledHeight, unsigned int vpW);
 
 	inline void loadATexture(string fileName, Texture *texture,vector<unsigned char> &image)
@@ -120,7 +120,7 @@ namespace Graphics
 	}
 
 	
-	inline void drawImage(Graphics::Texture *texture, unsigned int projectedX, unsigned int projectedY, float projectedWidth, float projectedHeight)
+	inline void drawImage(Graphics::Texture *texture, float projectedX, float projectedY, float projectedWidth, float projectedHeight)
 	{
 		glEnable(GL_TEXTURE_2D);
 		glBindTexture(GL_TEXTURE_2D, texture->id);
